refactoring : mypage 폴더구조 변경
