+++ src/app/mypage/edit/changepassword/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import CertifyPassword from '@/components/mypage/CertifyPassword';
-import NewPassword from '@/components/mypage/NewPassword';
-import PasswordChangeSuccess from '@/components/mypage/PasswordChangeSuccess';
+import CertifyPassword from '@/app/mypage/(mypage)/CertifyPassword';
+import NewPassword from '@/app/mypage/(mypage)/NewPassword';
+import PasswordChangeSuccess from '@/app/mypage/(mypage)/PasswordChangeSuccess';
 import { useAuthStore } from '@/store/userAuth.store';
 import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react';
@@ -19,7 +19,7 @@ const ChangePasswordPage = () => {
   }, [isLoggedIn, router]);
 
   return (
-    <div className='min-h-[calc(100dvh-134.5px)] flex flex-col sm:min-h-[calc(100vh-60px)] sm:mb-0 w-[90%] mx-auto sm:max-w-[640px]'>
+    <div className="min-h-[calc(100dvh-134.5px)] flex flex-col sm:min-h-[calc(100vh-60px)] sm:mb-0 w-[90%] mx-auto sm:max-w-[640px]">
       <div className="w-full mx-auto max-w-[340px] sm:max-w-[640px] sm:px-0">
         {step === 2 ? null : (
           <div onClick={() => router.back()} className="py-[14px]">
